refactor(mmpanelbase): moved code from header file into a seperate file
