+++ src/mmpanelbase.h
@@ -1,5 +1,6 @@
 /*******************************************************
  Copyright (C) 2006 Madhan Kanagavel
+ Copyright (C) 2015 James Higley
 
  This program is free software; you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by
@@ -24,106 +25,33 @@
 #include <wx/webviewfshandler.h>
 //----------------------------------------------------------------------------
 
-class wxSQLite3Database;
 class wxListItemAttr;
 
 class mmListCtrl: public wxListCtrl
 {
 public:
-    mmListCtrl(wxWindow *parent, wxWindowID winid)
-        : wxListCtrl(parent, winid, wxDefaultPosition, wxDefaultSize, wxLC_REPORT|wxLC_HRULES|wxLC_VRULES|wxLC_VIRTUAL|wxLC_SINGLE_SEL)
-        , attr1_(new wxListItemAttr(mmColors::listBorderColor, mmColors::listAlternativeColor0, wxNullFont))
-        , attr2_(new wxListItemAttr(mmColors::listBorderColor, mmColors::listAlternativeColor1, wxNullFont))
-        , m_selected_row(-1)
-        , m_selected_col(0)
-        , m_asc(true)
-    {}
-    virtual ~mmListCtrl()
-    {
-        if (attr1_) delete attr1_;
-        if (attr2_) delete attr2_;
-    }
-public:
+    mmListCtrl(wxWindow *parent, wxWindowID winid);
+    virtual ~mmListCtrl();
+
     wxListItemAttr *attr1_, *attr2_; // style1
     long m_selected_row;
     int m_selected_col;
     bool m_asc;
 
-public:
-    virtual wxListItemAttr* OnGetItemAttr(long row) const
-    {
-        return (row % 2) ? attr2_ : attr1_;
-    }
-    wxString BuildPage(const wxString &title) const
-    {
-        wxString text;
-        text << "<html>" << wxTextFile::GetEOL();
-        text << "<head>" << wxTextFile::GetEOL();
-        text << "<title>" << title << "</title>" << wxTextFile::GetEOL();
-        text << "</head>" << wxTextFile::GetEOL();
-        text << "<body>" << wxTextFile::GetEOL();
-        text << "<table ";
-        if ((GetWindowStyle() & wxLC_HRULES) ||
-          (GetWindowStyle() & wxLC_VRULES))
-            text << "border=1";
-        else
-            text << "border=0";
-        text << " cellpadding=4 cellspacing=0 >" << wxTextFile::GetEOL();
-
-        text << "<tr>" << wxTextFile::GetEOL();
-        for (int c = 0; c < GetColumnCount(); c++)
-        { 
-            wxListItem col;
-            col.SetMask(wxLIST_MASK_TEXT);
-            GetColumn(c, col);
-            text << "<th><i>" << col.GetText() << "</i></th>" << wxTextFile::GetEOL();
-        }
-        text << "</tr>" << wxTextFile::GetEOL();
-
-        for (int i = 0; i < GetItemCount(); i++)
-        { 
-            text << "<tr>" << wxTextFile::GetEOL();
-            for (int col = 0; col < GetColumnCount(); col++)
-            {
-                text << "<td>" << wxListCtrl::GetItemText(i, col) << "</td>" << wxTextFile::GetEOL();
-            }
-            text << "</tr>" << wxTextFile::GetEOL();
-        }
-        text << "</table>" << wxTextFile::GetEOL();
-        text << "</body>" << wxTextFile::GetEOL();
-        text << "</html>" << wxTextFile::GetEOL();
-
-        return text;
-    }
+    virtual wxListItemAttr* OnGetItemAttr(long row) const;
+    wxString BuildPage(const wxString &title) const;
 };
 
 class mmPanelBase : public wxPanel
 {
 public:
-    mmPanelBase() {}
-    virtual ~mmPanelBase() {}
+    mmPanelBase();
+    virtual ~mmPanelBase();
 
-public:
-    virtual wxString BuildPage() const { return "TBD"; }
-    virtual void PrintPage()
-    {
-        wxWebView * htmlWindow = wxWebView::New(this, wxID_ANY);
-        htmlWindow->RegisterHandler(wxSharedPtr<wxWebViewHandler>(new wxWebViewFSHandler("memory")));
-        htmlWindow->SetPage(BuildPage(), "");
-        htmlWindow->Print();
-    }
-    void windowsFreezeThaw()
-    {
-#ifdef __WXGTK__
-        return;
-#endif
+    virtual wxString BuildPage() const;
+    virtual void PrintPage();
+    void windowsFreezeThaw();
 
-        if (this->IsFrozen())
-            this->Thaw();
-        else
-            this->Freeze();
-    }
-public:
     virtual void sortTable() = 0;
 };
 //----------------------------------------------------------------------------
